requirements.txt
Flask==1.1.2
psycopg2-binary==2.8.6
peewee==3.14.1
marshmallow==3.10.0
